Moved out all of the tag code in event list to its own file
