+++ lib/pages/home_page.dart
@@ -2,12 +2,14 @@ import 'package:executive_planner/backend/event_list.dart';
 import 'package:executive_planner/backend/jason.dart';
 import 'package:executive_planner/backend/master_list.dart';
 import 'package:executive_planner/pages/calendar.dart';
-import 'package:executive_planner/pages/event_change_form.dart';
+import 'package:executive_planner/pages/forms/event_form.dart';
 import 'package:executive_planner/widgets/bottom_nav_bar.dart';
 import 'package:executive_planner/widgets/event_list_display.dart';
 import 'package:executive_planner/widgets/search.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'forms/event_form.dart';
 
 // TODO: Automatically hide unwanted events (subevents, trash, completed?)
 /// The starting page of the application.
@@ -135,7 +137,7 @@ class _ExecutiveHomePageState extends State<ExecutiveHomePage> {
     return Navigator.push(
       context,
       MaterialPageRoute(
-        builder: (context) => EventChangeForm(
+        builder: (context) => EventForm(
           event: event,
           events: widget.events,
         ),
